+++ R/convert_plink_binary_files_to_plink_text_files.R
@@ -1,0 +1,6 @@
+#' Convert PLINK binary files to PLINK plain-text files
+#' @export
+convert_plink_binary_files_to_plink_text_files <- function( # nolint indeed a long function name
+) {
+
+}